[frontend] eval form (#555)
Ok so this is the first sketch of the evaluation form of the circuit. It's really bare bones and really raw. But I am submitting it anyway so that we integrate it faster and avoid conflicts.

Right now, it does work and seems to be correct. But.

**It does not have the shape we want**. Right now the gates just emit bytecodes directly. This is not what we want for the purposes of constant propagation. Nor we can deduplicate committed internal witnesses based on this structure.

**The hint mechanism is unbaked**. It kind of gets into the shape envisioned but it's still not what we want. Specifically, right now it's still a gate and we don't want that. Moreover, we register the hint each time we emit such a gate. And finally, the serialization aspect is basically ignored. This is all left for the follow-up work.

**Non-optimal**. Right now this regressses witness filling performance from 30%-200% on smaller circuits (like sha256 2 KiB and zklogin). Curiously, it actually works 2x faster on bigger circuits like sha256 1 MiB. There are several reasons for that, like conditional branching on every load and store, and similiar stuff. Work on this is deferred for the follow-up PRs.

It's also concievable that this is not even getting back on par with the previous mechanism. That seems to be alright becuase it allows us to optimize the constraint system.