[frontend] Downgrade rsa test dependency to latest stable release
We use the rsa crate to create test data for the Rs256Verify circuit.

We used rsa-0.10.0-rc.* because these are the first versions with an API
that's compatible with rand-0.9 RNGs (and this is the version we use in
the monbijou project). The stable version 0.9.8 APIs are only compatible
with rand-0.6 RNGs.

The rsa-0.10.0-rc.* are currently broken because they depend on
crypto-bigint-0.7.0-rc.* and this crate introduced a breaking change. So
I think we should consider these rc releases to be unstable.

This PR solves this problem by downgrading to the latest stable rsa
release, 0.9.8. The dependency on rand-0.6 RNGs is avoided by including
a precomputed RSA private key for use in tests.