[frontend] rename internal to aux
Because this is a slightly different concept. To make it clear consider that the
outputs of the gates are represented by the internal wires. So in a sense those
wires that are created inside a gate they use internal wires but they are used
for a completely different thing.

Also add documentation.