[frontend] const eval
This changeset allows you to take a gate and evaluate it with the constant
inputs. This will be needed for the constant propagation pass.

This is a deliberate design decision to reuse the witness evaluation form
interpreter to avoid code duplication or introducing another way to compute the
values for the gate. I admit it's a bit clunky though. We could iterate and
improve it later.