[frontend] constant propagation
Optimization pass that takes all the defined constants and tries to evaluate
it with constants. If it succeeds, then the outputs of the gate can be replaced
with constants.

This is still behind a feature flag: to run it use the MONBIJOU_CONSTPROP=1
environment variable.

Note that while this performs replacement we don't prune the dead gates. This
is a task

https://linear.app/irreducible/issue/ENG2-15/opt-redundant-intermediate-wires