[frontend] fusion inline single user
We don't want to inline when any of the users cannot inline due to exceeding the
XOR terms budget (64), because that has no sense: we will grow the users but
won't eliminate the producer.

There is an exception though: there is only one user. In that case we should
inline into that user, but only if it is sound.