[examples] keccak: assert something
keccak prior to this changeset did not assert anything, it just computed hash
digest, that's all. Because the gates do not have any side effects, they only
produce some values, the whole thing is a subject for dead code elimination.

This changeset fixes this and constrains the final computed value to the
expected one.