[frontend] Make hash tweaking into a function (#819)
This PR changes the ChainTweak circuit from a struct:

```
pub struct ChainTweak {
  ... inout wires to be populated ...
}

impl ChainTweak {
  fn new(builder:: &CircuitBuilder, params) -> Self {
     ... build circuit ...
  }
}
```

To a function:

```
pub fn verify_chain_tweak(builder: &CircuitBuilder, params, ... inout wires ..., ) -> Keccak {
  ... build circuit ...
}
```

The `populate_*`functions are also removed, the caller is reponsible for populating the wires.

The function reutrns a Keccak verifier that must be populated with the expected digest and message.

A base `verify_tweaked_keccak`​ is also factored out which handles all tweaks of the form `Keccak256(param || tweak || additional)`​, this circuit will be used as a common function by all the tweaking circuits we need for hash-based signatures.

### Note

Ideally the message wries would be passed in as an argument to the function, but there’s added complexity with this approach. The Keccak circuit contains a padded version of the message. This would have to be constructed and passed into the `verify_chain_tweak`​ function separately. So in this case it is arguable that it’s simpler to reuse the Keccak struct.

​