more multiplexer tricks in sha (#877)
even more multiplexing tricks!

use yet another multiplexer in order to concoct the "expected" padding word, given the real padding word.

i did something similar when i rewrote keccak.

in the sha256 case, we have to be a bit tricky, because of the endianness of the high and low halves.