[sha256] sha256_fixed function for fixed-length inputs (#909)
This is useful for Bitcoin applications, like proving knowledge of an address private key. This involves a SHA-256 hash of a fixed-size payload.

I chose to make it the callers responsibility to split the wires into words, we'll need to add another function to help with that. Part of the rationale is to decouple concerns. More practically, the output digest will be the input to RIPEMD160 hash, which also operates on 32-bit words, so having interfaces for both that take inputs and outputs split into words is natural and efficient for chaining them.