[frontend] linear constraints
Introduce the concept of linear constraints into the constraint builder.

This is something we've discussed already and decided that it's a good idea. Our
constraint system does not support the linear constraints natively instead these
need to be transformed into AND constraints. That being said, there is still
value in having a higher-level representation because this would aid
optimizations like gate fusion.

Another change that is proposed here is modelling rotations. Those are not
supported by our constraint system either and there is also value in deferring
lowering them. Specifically this is useful to hold on for longer because we can
distribute them. In the lowered form (rotr(a, n) = (a >> n) | (a << (64-n)))
we cannot really do much about them.

As you could've guessed those changes are preparations for the new version of
gate fusion.