[frontend] gate fusion v2
Gate Fusion that works.

The previous approach was irreparably broken. It couldn't build a proper def
and use chains, couldn't inline things transitively. Now this implementation
comes with comprehensive testing.

This implementation gets its job done in three phases:

1. building an auxiliary structure called legraph (Linear Expression Graph)
2. deciding the committed set based on that graph
3. inlining

Legraph contains a graph which flows from the opaque values (inputs, values
defined by non-linear constraints, etc), through linear expressions defining
values and terminating in the non-linear expressions referred to as roots.

Then we traverse backwards from the non-linear expressions propagating
information/context through the edges making decisions whether to commit a value
or not. A value is committed if:

1. the linear expression cannot be inlined into at least one user.
2. the node has a certain maximum distance from a root. This is a proxy to gauge
   how many terms the inlined expression has and is used to prevent explosive
   behavior

Once we marked linear nodes which we have to commit we are going to perform
inlining. For each committed linear expression we are going to create a
corresponding AND constraint that is going to materialize that expression. Then
we aggressively inline every expression we see on the graph. This works because
at this point everything that should not be inlined is committed and as such
cannot be inlined. The same process is repeated for the existing AND constraints
(the roots).

Doing all of that we get patches that we can apply to the constraint system
builder (or validate).