[frontend] Use the Keccak::build_circuit API internally in tweaked hashers
This PR changes the circuit_tweaked_keccak to use the
Keccak::build_circuit API internally.

This is to demonstrate the
composition pattern of passing all 'external' wires (in this case the
digest and different forms of a message) into a function that builds a
circuit.

The current API of circuit_tweaked_keccak (i.e constructing and
returning a Keccak struct) is retained to avoid large changes to the
hash-based signature circuit at this stage.