[frontend] pinning wires
Gate fusion can aggressively inline every operator. Theoretically it might not
be the best decision for every case and directing gate fusion to not inline
a certain variable.

To help with that we introduce a special function `pin_wire` which allows the
circuit author to annotate the wires to be forced committed.