Gruen 3.2 memory savings (#1140)
This PR plays with the `multilinear::evaluate` trick of representing the eq indicator as an outer product of two smaller eq inds, and exploiting that in mlechecks. The primary target of this optimization is `SelectorMlecheck` in MUL reduction, which has to instantiate 64x equality indicators, but others benefit as well.

Pros:
 * The eq ind related memory consumption can be brought from `O(n)` to `O(sqrt(n))`
* Smaller chunk sizes pave the way to Four Russians style optimizations on binary mlechecks, similar to the ones AND reduction already employs

Cons:
 * This may be in the way of ZK sumcheck efforts (not knowledgeable enough on this)
 * This is a complication to the code (albeit small and limited scope)
* Sumcheck has to do additional `O(n/sqrt(n))` large field muls (though this is more than offset by the savings of not having to fully expand eq inds).

Apart from memory savings, benchmarking shows modest 2-5% perf improvement in single-threaded x86 MUL throughtput as well.