[shift] Refactor to avoid reading monster and public evals
This refactors the shift reduction prover, so that the only value read
by the verifier is the witness evaluation---the monster multilinear eval
and public eval are computed by the verifier. This is correct because
the verifier *can* compute them, whereas the witness eval is the claim
for the next reduction.